+++ Nos/Assets/Localization/Localized.swift
@@ -102,6 +102,8 @@ enum Localized: String, Localizable, CaseIterable {
     case done = "Done"
     case editProfile = "Edit Profile"
     case youNeedToEnterAPrivateKeyBeforePosting = "You need to enter a private key in Settings before you can publish posts."
+    case relayDoesNotSupportNIP40 = "This relay doesn't support expiring messages. Please select another one."
+    case anyRelaysSupportingNIP40 = "There aren't any relays that support expiring messages in your relay list. Please add one and retry."
     case nostrBuildHelp = "Use nostr.build to post image links."
     case noteDisappearsIn = "Note disappears in" // Choices like 15 mins, 1 hour, etc. appear after this text.
     case attachMedia = "Attach Media"