+++ Nos/Views/New Note/NewNoteView.swift
@@ -5,10 +5,11 @@
 //  Created by Matthew Lorentz on 2/6/23.
 //
 
-import SwiftUI
 import CoreData
-import SwiftUINavigation
 import Dependencies
+import Logger
+import SwiftUI
+import SwiftUINavigation
 
 struct NewNoteView: View {
     
@@ -101,7 +102,7 @@ struct NewNoteView: View {
     }
 
     private func publishPost() async {
-        guard let keyPair = currentUser.keyPair else {
+        guard let keyPair = currentUser.keyPair, let author = currentUser.author else {
             alert = AlertState(title: {
                 TextState(Localized.error.string)
             }, message: {
@@ -116,24 +117,30 @@ struct NewNoteView: View {
             if let expirationTime {
                 tags.append(["expiration", String(Date.now.timeIntervalSince1970 + expirationTime)])
             }
+
+            let jsonEvent = JSONEvent(pubKey: keyPair.publicKeyHex, kind: .text, tags: tags, content: content)
             
-            let jsonEvent = JSONEvent(
-                id: "",
-                pubKey: keyPair.publicKeyHex,
-                createdAt: Int64(Date().timeIntervalSince1970),
-                kind: 1,
-                tags: tags,
-                content: content,
-                signature: ""
-            )
-            
-            if let selectedRelay {
-                try await relayService.publish(
-                    event: jsonEvent,
-                    to: selectedRelay,
-                    signingKey: keyPair,
-                    context: viewContext
-                )
+            if let relay = selectedRelay {
+                guard expirationTime == nil || relay.metadata?.supportedNIPs?.contains(40) ?? false else {
+                    alert = AlertState(title: {
+                        TextState(Localized.error.string)
+                    }, message: {
+                        TextState(Localized.relayDoesNotSupportNIP40.string)
+                    })
+                    return
+                }
+                try await relayService.publish(event: jsonEvent, to: relay, signingKey: keyPair, context: viewContext)
+            } else if expirationTime != nil {
+                let relays = try await Relay.find(supporting: 40, for: author, context: viewContext)
+                guard !relays.isEmpty else {
+                    alert = AlertState(title: {
+                        TextState(Localized.error.string)
+                    }, message: {
+                        TextState(Localized.anyRelaysSupportingNIP40.string)
+                    })
+                    return
+                }
+                try await relayService.publish(event: jsonEvent, to: relays, signingKey: keyPair, context: viewContext)
             } else {
                 try await relayService.publishToAll(event: jsonEvent, signingKey: keyPair, context: viewContext)
             }
@@ -147,11 +154,7 @@ struct NewNoteView: View {
             }, message: {
                 TextState(error.localizedDescription)
             })
-            // Replace this implementation with code to handle the error appropriately.
-            // fatalError() causes the application to generate a crash log and terminate. You should not use this
-            // function in a shipping application, although it may be useful during development.
-            let nsError = error as NSError
-            fatalError("Unresolved error \(nsError), \(nsError.userInfo)")
+            Log.error(error.localizedDescription)
         }
     }
 }